#29. Tạo Page by id | Next.JS Cơ Bản Cho Beginners với React và Typescript
