+++ HoiDanIT/learn-nextjs/next13-starter-typescript/src/app/blogs/[id]/page.tsx
@@ -1,0 +1,7 @@
+const ViewDetailBlog = ({ params }: { params: { id: string } }) => {
+  console.log('>>> check props: ', params.id)
+
+  return <div>View Detial Blog With Duy Nghia Dev = {params.id}</div>
+}
+
+export default ViewDetailBlog
