+++ HoiDanIT/learn-nextjs/next13-starter-typescript/src/app/page.tsx
@@ -2,25 +2,10 @@
 import Link from 'next/link'
 import x from '@/styles/app.module.css'
 import y from '@/styles/hoidanit.module.css'
-import AppTable from '@/components/app.table'
-import useSWR from 'swr'
 
 export default function Home() {
-  const fetcher = (url: string) => fetch(url).then((res) => res.json())
-
-  const { data, error, isLoading } = useSWR('http://localhost:8000/blogs', fetcher, {
-    revalidateIfStale: false,
-    revalidateOnFocus: false,
-    revalidateOnReconnect: false
-  })
-
-  if (!data) {
-    return <div>Loading...</div>
-  }
-
   return (
     <div>
-      <div>{data?.length}</div>
       <ul>
         <li className={x['red']}>
           <Link href={'/facebook'}>
@@ -34,7 +19,6 @@ export default function Home() {
           <Link href={'/tiktok'}>Tiktok</Link>
         </li>
       </ul>
-      <AppTable blogs={data?.sort((a: any, b: any) => b.id - a.id)} />
     </div>
   )
 }
