+++ HoiDanIT/learn-nextjs/next13-starter-typescript/src/components/app.header.tsx
@@ -16,14 +16,8 @@ const AppHeader = () => {
         <Navbar.Toggle aria-controls='basic-navbar-nav' />
         <Navbar.Collapse id='basic-navbar-nav'>
           <Nav className='me-auto'>
-            <Link href={'/facebook'} className='nav-link'>
-              Facebook
-            </Link>
-            <Link href={'/tiktok'} className='nav-link'>
-              Tiktok
-            </Link>
-            <Link href={'/youtube'} className='nav-link'>
-              Youtube
+            <Link href={'/blogs'} className='nav-link'>
+              Blogs
             </Link>
           </Nav>
         </Navbar.Collapse>
