+++ HoiDanIT/learn-nextjs/next13-starter-typescript/src/components/app.table.tsx
@@ -4,6 +4,7 @@ import { Button } from 'react-bootstrap'
 import CreateModal from './create.modal'
 import { useState } from 'react'
 import UpdateModal from './update.modal'
+import Link from 'next/link'
 
 interface IProps {
   blogs: IBlog[]
@@ -41,7 +42,9 @@ const AppTable = (props: IProps) => {
                 <td>{item.title}</td>
                 <td>{item.author}</td>
                 <td>
-                  <Button>View</Button>
+                  <Link className='btn btn-primary' href={`/blogs/${item.id}`}>
+                    View
+                  </Link>
                   <Button
                     variant='warning'
                     className='mx-3'
